+++ koru-compiler-plugin/src/main/kotlin/com/futuremind/koru/gradle/CompilerPlugin.kt
@@ -30,7 +30,7 @@ class CompilerPlugin : Plugin<Project> {
 
     private fun Project.enableKspRunForCommonMainSourceSet() = dependencies {
         //todo don't hardcode version
-        add("kspCommonMainMetadata", "com.futuremind:koru-processor:0.11.1")
+        add("kspCommonMainMetadata", "com.futuremind:koru-processor:0.12.0")
     }
 
     private fun Project.makeSureCompilationIsRunAfterKsp() = tasks