+++ koru-processor/src/jvmMain/kotlin/com/futuremind/koru/processor/builders/WrapperInterfaceBuilder.kt
@@ -12,7 +12,7 @@ class WrapperInterfaceBuilder(
 ) : WrapperBuilder(originalTypeName, originalTypeSpec, generatedInterfaces) {
 
     private val functions = originalTypeSpec.funSpecs
-        .filter { !it.modifiers.contains(KModifier.PRIVATE) }
+        .filter { !it.modifiers.isPrivateOrProtected() }
         .map { originalFuncSpec ->
             originalFuncSpec.toBuilder(name = originalFuncSpec.name)
                 .clearBody()
@@ -20,13 +20,15 @@ class WrapperInterfaceBuilder(
                 .apply {
                     modifiers.add(KModifier.ABSTRACT)
                     modifiers.remove(KModifier.SUSPEND)
+                    modifiers.remove(KModifier.ACTUAL)
+                    modifiers.remove(KModifier.EXPECT)
                 }
                 .setupOverrideModifier(originalFuncSpec)
                 .build()
         }
 
     private val properties = originalTypeSpec.propertySpecs
-        .filter { !it.modifiers.contains(KModifier.PRIVATE) }
+        .filter { !it.modifiers.isPrivateOrProtected() }
         .map { originalPropertySpec ->
             PropertySpec
                 .builder(
@@ -34,7 +36,11 @@ class WrapperInterfaceBuilder(
                     type = originalPropertySpec.wrappedType
                 )
                 .mutable(false)
-                .apply { modifiers.add(KModifier.ABSTRACT) }
+                .apply {
+                    modifiers.add(KModifier.ABSTRACT)
+                    modifiers.remove(KModifier.ACTUAL)
+                    modifiers.remove(KModifier.EXPECT)
+                }
                 .setupOverrideModifier(originalPropertySpec)
                 .build()
         }